requirements.txt
altgraph==0.17
appdirs==1.4.4
asgiref==3.3.4
backcall==0.2.0
chardet==4.0.0
click==7.1.2
decorator==4.4.2
distlib==0.3.1
docutils==0.16
et-xmlfile==1.0.1
filelock==3.0.12
Flask==1.1.2
gunicorn==20.1.0
idna==2.10
ipykernel==5.5.0
ipython==7.20.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.18.0
Jinja2==2.11.3
lxml==4.6.2
MarkupSafe==1.1.1
numpy==1.20.1
packaging==20.9
pandas==1.2.2
pandas-datareader==0.9.0
parso==0.8.1
pefile==2019.4.18
pickleshare==0.7.5
Pillow==8.1.1
prompt-toolkit==3.0.16
Pygments==2.8.0
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.1
six==1.15.0
typing-extensions==3.7.4.3
virtualenv==20.4.3
wcwidth==0.2.5
Werkzeug==1.0.1
xlrd==2.0.1